Rename: data 폴더 분리 data 폴더 분리, 컴포넌트 파일 정리
